feat: bank transactions ui iteration
